requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
face_recognition==1.3.0
qdrant-client==1.6.4
python-multipart==0.0.6
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
jinja2==3.1.2 